I broke login. 1. Broken login 2. Make login and registration components
